refactor(cli): 配信元 / デフォルト ref などの内部定数を constants.ts に集約
各コマンドに散らばっていた以下を packages/lism-cli/src/constants.ts に集約：

- SOURCE_REPO / RAW_GITHUB_BASE
- DEFAULT_UI_REF / DEFAULT_SKILL_REF / DEFAULT_TEMPLATES_REF
- UI_COMPONENTS_PATH / UI_HELPER_PATH / UI_REGISTRY_INDEX_PATH
- SKILL_SOURCE_PATH / EXAMPLES_PATH

dev / main マージ前に書き換える ref の FIXME も 1 箇所にまとめ、
マージ運用時の作業が constants.ts のみで完結するようにした。
